+++ python/tvm/relay/frontend/pytorch.py
@@ -3845,6 +3845,33 @@ def get_all_op_names(graph):
     return set(node.kind() for node in nodes)
 
 
+def inline_qparams(graph, param_tensors):
+    import torch
+    getattr_nodes = graph.findAllNodes("prim::GetAttr", recurse=True)
+    seen = set()
+    for node in getattr_nodes:
+        out_name = node.output().debugName()
+
+        if "_input_scale" in out_name:
+            out_scale_node = graph.create("prim::Constant")
+            out_scale_node.insertBefore(node)
+            for k, v in param_tensors.items():
+                if out_name in k:
+                    out_scale_node.f_("value", v.numpy())
+                    break
+            out_scale_node.output().setType(torch._C.FloatType.get())
+            node.replaceAllUsesWith(out_scale_node)
+        elif "_input_zero_point" in out_name:
+            out_zero_point_node = graph.create("prim::Constant")
+            out_zero_point_node.insertBefore(node)
+            for k, v in param_tensors.items():
+                if out_name in k:
+                    out_zero_point_node.i_("value", v.numpy().item())
+                    break
+            out_zero_point_node.output().setType(torch._C.IntType.get())
+            node.replaceAllUsesWith(out_zero_point_node)
+
+
 def from_pytorch(
     script_module,
     input_infos,
@@ -3911,6 +3938,7 @@ def from_pytorch(
 
     graph = script_module.graph.copy()
     _run_jit_passes(graph)
+    print(graph)
 
     if custom_convert_map:
         converter.update_convert_map(custom_convert_map)
@@ -3941,7 +3969,7 @@ def from_pytorch(
         weight_quant_params = qnn_torch.get_weight_quant_params(
             script_module, packed_param_map.values()
         )
-        qnn_torch.inline_qparams(graph, tensors)
+        inline_qparams(graph, tensors)
         input_scales_for_bias = qnn_torch.add_input_quant_params_to_op_inputs(graph)
         qnn_torch.add_quant_params_to_outputs(
             outputs,
